# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenClaw Mail Secretary API
  version: 2.0.0
  description: >
    Canonical HTTPS REST API surface for clawd-on-desk Electron HUD,
    openclaw-hud-gateway, and the private organization-scoped OpenClaw Cloud Run service.
    The HUD uses only the IAP-protected gateway URL; it never calls the private
    openclaw-api URL directly. Initial OAuth is browser navigation through the
    gateway. The private API performs confidential Google OAuth code exchange
    with the OAuth client secret from Secret Manager, stores the Gmail refresh
    token in Secret Manager, and returns only short-lived Google ID tokens plus
    sanitized claims to the HUD. The HUD never receives, sends, or persists
    refresh tokens, access tokens, authorization codes, code verifiers, or
    OAuth client secrets.
servers:
  - url: https://{gatewayHost}
    variables:
      gatewayHost:
        default: openclaw-hud-gateway.example.run.app
security:
  - googleIdToken: []
...
paths:
...
  /v1/auth/start:
    get:
      tags: [auth]
      summary: Start browser OAuth through the IAP-protected HUD gateway.
      description: >
        Browser navigation entrypoint used by the HUD for initial login.
        The request is not a programmatic Bearer-token API call; it is protected
        by the user's IAP browser session on the gateway. The gateway asks the
        private API to create a short-lived OAuth auth session, then redirects
        the browser to Google's authorization URL. The HUD supplies only a local
        loopback return_uri. The HUD does not generate or store the OAuth
        code_verifier and never receives the authorization code. It generates
        a separate random handoff_verifier and sends only its S256 challenge.
      operationId: startOAuth
      security: []
      parameters:
        - name: return_uri
          in: query
          required: true
          description: >
            Exact HUD loopback receiver. Only numeric loopback hosts, ports
            1024-65535, and path /callback are accepted. Userinfo, query,
            fragment, localhost aliases, and external hosts are rejected.
          schema:
            type: string
            format: uri
            pattern: '^http://(127\.0\.0\.1|\[::1\]):(102[4-9]|10[3-9][0-9]|1[1-9][0-9]{2}|[2-9][0-9]{3}|[1-5][0-9]{4}|6[0-4][0-9]{3}|65[0-4][0-9]{2}|655[0-2][0-9]|6553[0-5])/callback$'
            example: "http://127.0.0.1:23456/callback"
        - name: handoff_challenge
          in: query
          required: true
          description: Base64url SHA-256 challenge for the HUD-held handoff_verifier.
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{43}$'
      responses:
        "302":
          description: Redirect to Google authorization URL.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/auth/callback:
    get:
      tags: [auth]
      summary: Google OAuth redirect URI handled by the HUD gateway.
      description: >
        Google redirects the browser here with code and state. The gateway
        forwards code and state to the private API internal exchange endpoint.
        The private API validates state, performs confidential token exchange
        using the OAuth client secret from Secret Manager, verifies ID token
        claims and stores the Gmail refresh token in Secret Manager. The
        gateway returns a no-store transfer page that POSTs only a random
        single-use handoffCode to the HUD loopback return_uri, never an ID
        token or claims. The auth session binds return_uri, state, nonce, OAuth
        PKCE verifier, and HUD handoff challenge; the code expires after 60 seconds.
        The page uses Cache-Control: no-store, Referrer-Policy: no-referrer, and
        a script-src 'nonce-*'; default-src 'none' CSP. No refresh token is
        returned to the HUD.
      operationId: oauthCallback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: No-store browser handoff page for the HUD loopback listener.
          content:
            text/html:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/auth/handoff/redeem:
    post:
      tags: [auth]
      summary: Redeem the loopback handoff code using the HUD-held verifier.
      description: >
        Pre-login endpoint hosted by the narrowly exposed auth-bootstrap
        service. The HUD main process submits the single-use handoffCode and
        original high-entropy handoffVerifier. The bootstrap service forwards
        them using Bootstrap SA OIDC. The private API atomically consumes the
        code, compares S256(verifier) to the bound challenge, and only then
        returns a short-lived ID token. A lost response requires a new OAuth
        flow; a consumed code is never made reusable.
      operationId: redeemAuthHandoff
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoffRedeemRequest"
      responses:
        "200":
          description: Single-use handoff redeemed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthHandoffResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /internal/auth/oauth-sessions:
    post:
      tags: [internal]
      summary: Create a short-lived OAuth auth session for the gateway.
      description: >
        Gateway-to-private-API internal endpoint. The private API creates
        state and code_verifier, stores them in a short-lived auth session, and
        binds the exact return URI and HUD-provided handoff challenge before
        returning a Google authorization URL. This endpoint is not directly
        callable by the HUD.
      operationId: createOAuthSession
      security:
        - gatewayOidcToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthSessionCreateRequest"
      responses:
        "200":
          description: OAuth auth session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthSessionCreateResponse"
        "401":
          $ref: "#/components/responses/InternalUnauthorized"
        "403":
          $ref: "#/components/responses/InternalForbidden"
  /internal/auth/oauth-exchange:
    post:
      tags: [internal]
      summary: Confidentially exchange Google OAuth code and store refresh token.
      description: >
        Gateway-to-private-API internal endpoint. The private API validates
        state, loads the stored code_verifier, exchanges the authorization code
        at Google's token endpoint using the OAuth client secret from Secret
        Manager, validates ID token claims, stores the Gmail refresh token in
        Secret Manager, stores the short-lived ID token behind the bound
        handoff challenge, and returns only a random one-time handoffCode.
      operationId: exchangeOAuthCode
      security:
        - gatewayOidcToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthExchangeRequest"
      responses:
        "200":
          description: OAuth exchange completed and refresh token stored server-side.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthExchangeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InternalUnauthorized"
        "403":
          $ref: "#/components/responses/InternalForbidden"
  /internal/auth/handoff/redeem:
    post:
      tags: [internal]
      summary: Atomically redeem a handoff code. Invoked only by the gateway.
      operationId: redeemInternalAuthHandoff
      security:
        - bootstrapOidcToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoffRedeemRequest"
      responses:
        "200":
          description: Handoff consumed and ID token returned to the gateway.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthHandoffResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InternalUnauthorized"
        "403":
          $ref: "#/components/responses/InternalForbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    googleIdToken:
      type: http
      scheme: bearer
      bearerFormat: Google ID token
      description: >
        Google-signed ID token returned to the HUD after gateway/API
        confidential OAuth exchange. Sent by the HUD to gateway-facing APIs in
        Authorization, and also in Proxy-Authorization for the first IAP
        programmatic-access approach. The private OpenClaw application validates
        issuer, expiration, aud (the OAuth client ID), hosted domain, and
        allowed subject. Cloud Run URL audience is used only for gateway,
        Scheduler, and Tasks OIDC tokens.
    gatewayOidcToken:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC token
      description: >
        OIDC identity token minted by the openclaw-hud-gateway service account
        for private openclaw-api internal calls. The gateway service account
        must have Cloud Run roles/run.invoker on openclaw-api. The private API
        validates issuer, expiration, audience equal to the openclaw-api Cloud
        Run service URL, and service account email allowlist.
    bootstrapOidcToken:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC token
      description: >
        OIDC identity token minted by the auth-bootstrap service account. The
        private API accepts it only on the internal handoff redeem endpoint.
    adminOidcToken:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC token
      description: >
        OIDC identity token minted by the dedicated admin service account.
        The private API validates issuer, expiration, audience equal to the
        openclaw-api Cloud Run service URL, and exact Admin SA email allowlist.
        Gateway, Scheduler, Tasks, runtime, and end-user tokens are rejected.
    schedulerOidcToken:
      type: http
...
  schemas:
...
    SanitizedAuthClaims:
      type: object
      additionalProperties: false
      properties:
        sub:
          type: string
        email:
          type: string
          format: email
        hd:
          type: string
        aud:
          type: string
        exp:
          type: integer
    BootstrapState:
      type: object
      required: [completed, bindingHash, reason]
      additionalProperties: false
      properties:
        completed:
          type: boolean
          description: True when backend confidential exchange has stored the Gmail refresh token in Secret Manager for this subject.
        bindingHash:
          type: [string, "null"]
          description: SHA-256 over gateway URL, organization domain, OAuth client ID, and Google subject. Non-secret local metadata for the HUD.
        reason:
          type: [string, "null"]
          enum: [missing, invalid, completed, null]
    AuthHandoffResponse:
      type: object
      required: [idToken, expiresAt, claims, bootstrap]
      additionalProperties: false
      properties:
        idToken:
          type: string
          description: Short-lived Google ID token. Never log and never expose to renderer IPC payloads.
        expiresAt:
          type: string
          format: date-time
        claims:
          $ref: "#/components/schemas/SanitizedAuthClaims"
        bootstrap:
          $ref: "#/components/schemas/BootstrapState"
    AuthExchangeResponse:
      type: object
      required: [handoffCode, expiresAt, returnUri]
      additionalProperties: false
      properties:
        handoffCode:
          type: string
          minLength: 43
          description: Random single-use code. It contains no token or claims and expires after 60 seconds.
        expiresAt:
          type: string
          format: date-time
        returnUri:
          type: string
          format: uri
    HandoffRedeemRequest:
      type: object
      required: [handoffCode, handoffVerifier]
      additionalProperties: false
      properties:
        handoffCode:
          type: string
          minLength: 43
        handoffVerifier:
          type: string
          minLength: 43
          maxLength: 128
          pattern: '^[A-Za-z0-9._~-]+$'
...
    OAuthSessionCreateRequest:
      type: object
      required: [returnUri, handoffChallenge]
      additionalProperties: false
      properties:
        returnUri:
          type: string
          format: uri
          pattern: '^http://(127\.0\.0\.1|\[::1\]):(102[4-9]|10[3-9][0-9]|1[1-9][0-9]{2}|[2-9][0-9]{3}|[1-5][0-9]{4}|6[0-4][0-9]{3}|65[0-4][0-9]{2}|655[0-2][0-9]|6553[0-5])/callback$'
          description: HUD loopback URI supplied to gateway /v1/auth/start.
        handoffChallenge:
          type: string
          pattern: '^[A-Za-z0-9_-]{43}$'
          description: S256 challenge for a verifier that remains only in the HUD main process.
    OAuthSessionCreateResponse:
      type: object
      required: [authorizationUrl, state, expiresAt]
      additionalProperties: false
      properties:
        authorizationUrl:
          type: string
          format: uri
        state:
          type: string
        expiresAt:
          type: string
          format: date-time
    OAuthExchangeRequest:
      type: object
      required: [code, state]
      additionalProperties: false
      properties:
        code:
          type: string
        state:
          type: string
...
    ErrorResponse:
      type: object
      required: [code, message, correlationId]
      additionalProperties: false
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          maxLength: 200
          description: Safe user-facing message only. Raw exception text, stack traces, HTTP headers, request/response bodies, tokens, API keys, secret resource names, and PII are prohibited.
        correlationId:
          type: string
        details:
          description: Sanitized structured validation details. Never contains request values or PII.
          type: array
          items:
            type: object
            required: [path, keyword, message, params]
            additionalProperties: false
            properties:
              path:
                type: string
              keyword:
                type: string
              message:
                type: string
              params:
                type: object
                additionalProperties: true
        supersededByProposalId:
          type: [string, "null"]
          description: >
            Present only when code is proposal_superseded. The current proposal
            revision the HUD should switch to. Null when the proposal was
            cancelled rather than superseded (D-35).
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - validation_error
        - unauthorized
        - forbidden
        - stale_card_version
        - invalid_approval_token
        - rate_limited
        - bootstrap_required
        - oauth_session_expired
        - oauth_exchange_failed
        - refresh_token_missing
        - refresh_token_invalid
        - operation_conflict
        - replan_required
        - external_result_unknown
        - proposal_superseded
        - pii_review_required
        - endpoint_not_configured
        - db_unavailable
        - service_unavailable
        - internal_error
        - rate_limit_exceeded  # Backward-compatible alias. Prefer rate_limited for MVP 429 responses.
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid Google ID/OIDC token. The HUD must not retry. OpenClaw returns ErrorResponse when application-level token validation fails.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorized by hosted domain, OAuth client ID, subject, or service-account allowlist. The HUD must not retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalUnauthorized:
      description: Missing or invalid OIDC token for an internal Cloud Scheduler / Cloud Tasks caller. Treat as IAM/OIDC configuration failure and alert; this is not a HUD retry condition.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalForbidden:
      description: Internal caller authenticated but not authorized. Treat as IAM binding failure and alert; this is not a HUD retry condition.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: "Stale cardVersion (code: stale_card_version), invalid or expired approvalToken (code: invalid_approval_token), or operation conflict (code: operation_conflict)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded. HUD should back off using Retry-After and must not auto-retry side-effect operations without refetching current card state.
      headers:
        Retry-After:
          description: Required. Seconds until this operation may be retried.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                required: [code]
                properties:
                  code:
                    const: rate_limited
          examples:
            rateLimited:
              value:
                code: rate_limited
                message: "Rate limit exceeded. Retry after the indicated delay."
                correlationId: "00000000-0000-0000-0000-000000000000"
...